requirements.txt
Package                      Version
---------------------------- -----------
aiohappyeyeballs             2.4.3
aiohttp                      3.10.8
aiosignal                    1.3.1
annotated-types              0.7.0
anyio                        4.6.0
attrs                        24.2.0
blinker                      1.8.2
cachetools                   5.5.0
certifi                      2024.8.30
charset-normalizer           3.3.2
click                        8.1.7
colorama                     0.4.6
deprecation                  2.1.0
Flask                        3.0.3
frozenlist                   1.4.1
google-ai-generativelanguage 0.6.10
google-api-core              2.20.0
google-api-python-client     2.147.0
google-auth                  2.35.0
google-auth-httplib2         0.2.0
google-generativeai          0.8.2
googleapis-common-protos     1.65.0
gotrue                       2.9.1
grpcio                       1.66.2
grpcio-status                1.66.2
h11                          0.14.0
h2                           4.1.0
hpack                        4.0.0
httpcore                     1.0.5
httplib2                     0.22.0
httpx                        0.27.2
hyperframe                   6.0.1
idna                         3.10
itsdangerous                 2.2.0
Jinja2                       3.1.4
MarkupSafe                   2.1.5
multidict                    6.1.0
packaging                    24.1
pillow                       10.4.0
pip                          24.2
postgrest                    0.17.0
proto-plus                   1.24.0
protobuf                     5.28.2
pyasn1                       0.6.1
pyasn1_modules               0.4.1
pydantic                     2.9.2
pydantic_core                2.23.4
pyparsing                    3.1.4
python-dateutil              2.9.0.post0
python-decouple              3.8
realtime                     2.0.5
requests                     2.32.3
rsa                          4.9
setuptools                   73.0.1
six                          1.16.0
sniffio                      1.3.1
storage3                     0.8.0
StrEnum                      0.4.15
supabase                     2.8.1
supafunc                     0.6.0
tqdm                         4.66.5
typing_extensions            4.12.2
uritemplate                  4.1.1
urllib3                      2.2.3
websockets                   13.1
Werkzeug                     3.0.4
wheel                        0.44.0
yarl                         1.13.1
